# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PedagangPulsa API
  description: |
    PedagangPulsa Member Mobile API v1.0

    ## Authentication

    The API uses JWT (JSON Web Token) based authentication with a 2-step verification process:

    1. **Login** → Get Access Token (15min) + Refresh Token (7days)
    2. **PIN Verify** → Get PIN Session Token (5min) for sensitive operations

    ## Idempotency

    Transaction and Transfer endpoints require `X-Reference-Id` header to prevent duplicate requests.

  version: 1.0.0
  contact:
    name: PedagangPulsa API Support
    email: api@pedagangpulsa.com
    url: https://pedagangpulsa.com
  license:
    name: Proprietary
servers:
  - url: https://api.pedagangpulsa.com/api/v1
    description: Production server
  - url: https://sandbox-api.pedagangpulsa.com/api/v1
    description: Sandbox server
...
paths:
...
  /topup:
    post:
      tags:
        - Topup
      summary: Request topup
      description: Submit topup request with proof of transfer
      operationId: requestTopup
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TopupRequest'
      responses:
        '201':
          description: Topup request created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TopupResponse'
  /topup/history:
    get:
      tags:
        - Topup
      summary: Get topup history
      description: Get user's topup request history
      operationId: getTopupHistory
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        '200':
          description: Topup history retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TopupHistoryResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login
...
  schemas:
    SuccessResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total_items:
          type: integer
          example: 150
        total_pages:
          type: integer
          example: 8

    # Auth schemas
...
    TopupRequest:
      type: object
      required:
        - amount
        - bank_code
        - transfer_proof
      properties:
        amount:
          type: number
          format: decimal
          minimum: 10000
          maximum: 10000000
        bank_code:
          type: string
        transfer_proof:
          type: string
          format: binary
        notes:
          type: string
          maxLength: 500
    TopupResponse:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        bank_code:
          type: string
        transfer_proof_url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
    TopupHistoryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TopupResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Product schemas
...